banners

+++ public/js/backend.js
@@ -4,9 +4,30 @@ function delUser(id)
         return
 
     fetch('/admin/users/delete/' + id, {
-        method: 'delete'
+        method: 'delete',
+        /*body: JSON.stringify({
+            'name': 'zzz'
+        }),*/
+        credentials: "same-origin"
     }).then(function (response) {
-        console.log(response)
+        //console.log(response)
+        if(response.status == 200) window.location.reload()
+    })
+}
+
+function delBanner(id)
+{
+    if(!(confirm('Are you sure?')))
+        return
+
+    fetch('/admin/banner/delete/' + id, {
+        method: 'delete',
+        /*body: JSON.stringify({
+            'name': 'zzz'
+        }),*/
+        credentials: "same-origin"
+    }).then(function (response) {
+        //console.log(response)
         if(response.status == 200) window.location.reload()
     })
 }
